+++ sell/sell_domain/src/main/java/com/sammy/sell_domain/data/SearchCarResult.kt
@@ -13,7 +13,7 @@ data class SearchCarResult(
     val imageUrl: String,
     val installment: Int,
     val licensePlate: String,
-    val loanValue: Int,
+    val loanValue: Double,
     val marketplaceOldPrice: Int,
     val marketplacePrice: Int,
     val marketplaceVisibleDate: String,
